Refactor: Strategy 클래스 네임 변경 및 Newsletter 도메인 패키지 구조 수정
