+++ src/main/java/com/newzet/api/newsletter/domain/model/NewsletterStatus.java
@@ -1,4 +1,4 @@
-package com.newzet.api.newsletter.domain;
+package com.newzet.api.newsletter.domain.model;
 
 public enum NewsletterStatus {
 	REGISTERED,
