+++ src/main/java/com/newzet/api/newsletter/domain/strategy/RandomRecommendationStrategy.java
@@ -1,4 +1,4 @@
-package com.newzet.api.newsletter.recommend;
+package com.newzet.api.newsletter.domain.strategy;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -6,10 +6,10 @@ import java.util.List;
 
 import org.springframework.stereotype.Component;
 
-import com.newzet.api.newsletter.domain.Newsletter;
+import com.newzet.api.newsletter.domain.model.Newsletter;
 
 @Component
-public class RandomRecommendationFactory implements RecommendationFactory {
+public class RandomRecommendationStrategy implements RecommendationStrategy {
 
 	public List<Newsletter> createRecommendationList(List<Newsletter> candidateList,
 		int count) {
