+++ src/main/java/com/newzet/api/newsletter/domain/strategy/RecommendationStrategy.java
@@ -1,0 +1,10 @@
+package com.newzet.api.newsletter.domain.strategy;
+
+import java.util.List;
+
+import com.newzet.api.newsletter.domain.model.Newsletter;
+
+public interface RecommendationStrategy {
+
+	List<Newsletter> createRecommendationList(List<Newsletter> candidateList, int count);
+}
