+++ src/test/java/com/newzet/api/newsletter/fixture/NewsletterFixture.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 import com.newzet.api.category.domain.Category;
 import com.newzet.api.category.repository.CategoryEntity;
 import com.newzet.api.newsletter.business.dto.NewsletterCacheDto;
-import com.newzet.api.newsletter.domain.Color;
-import com.newzet.api.newsletter.domain.Newsletter;
+import com.newzet.api.newsletter.domain.model.Color;
+import com.newzet.api.newsletter.domain.model.Newsletter;
 import com.newzet.api.newsletter.repository.NewsletterEntity;
 
 public class NewsletterFixture {
